Added a couple things. Mostly added further comments This was to help out in future should I want to change anything, Look at me being a nice coder to myself. I don't usually do that.
Started adding the command and function to start the rust server.

Can be reused for any other type of server.
New batches just be made for it then it can be repurposed.
Killing the other processes could be awkward but we'll wait and see.

A little bit of cleaning up on this too, somethings commented out.

Will look into removing some of the unused code later down the line.